backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-environ==0.12.0
django-cors-headers==4.6.0
gunicorn==23.0.0
psycopg[binary]==3.2.12
